Add some functions including change preview size and focus size (by press menu button), fixed landscape mode and take snapshot (by touch the screen).
We extend the JavaCameraView class to create a new View, where we add
our own function (e.g., get a list of preview sizes by using the
protected member mCamera). 

We use the APIs of android "mCamera.setDisplayOrientation()" to flip the
image, instead of the OpenCV one, which is faster.

We also fix the app to landscape mode.

We also allow the preview size and focus size changed by press menu
button. You can also take a snapshoot by touching the screen.

However manu focus is still not working. When we try to set the
parameter to the camera, it always fail.